database_microservice/requirements.txt
dj-database-url==1.3.0
Django==3.2.18
djangorestframework==3.14.0
psycopg2-binary==2.9.6
PyYAML==6.0
gunicorn==20.1.0
drf-yasg==1.21.5